move se_pair_workflow, fieldmap_to_phasediff; really add tests
